refactor: move training loops to module
